requirements.txt
# Kafka clients
kafka-python==2.0.2
confluent-kafka==2.3.0

# Data processing
pandas==2.1.4
numpy==1.26.2
pyspark==3.5.0

# API clients
requests==2.31.0
python-dotenv==1.0.0

# Database connectors
snowflake-connector-python==3.6.0
snowflake-sqlalchemy==1.5.1

# Monitoring and logging
prometheus-client==0.19.0
psutil==5.9.6

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0

# Utilities
pytz==2023.3